Extract the dialog into its own separate component/plugin that attaches itself to Mirador's Window component.
