+++ __tests__/miradorDownloadPlugin.test.js
@@ -5,15 +5,12 @@ import miradorDownloadPlugin from '../src';
 function createWrapper(props) {
   return shallow(
     <miradorDownloadPlugin.component
-      canvasLabel={label => (label || 'My Canvas Title')}
-      canvases={[]}
-      classes={{}}
-      manifest={{ getSequences: () => [] }}
-      viewType="single"
-      windowId="wid123"
+      handleClose={() => {}}
+      openDownloadDialog={() => {}}
+
       {...props}
     />,
-  ).dive();
+  );
 }
 
 describe('miradorDownloadPlugin', () => {
@@ -28,60 +25,13 @@ describe('miradorDownloadPlugin', () => {
   });
 
   describe('MenuItem', () => {
-    it('udpates the modalDisplayed state which clicked', () => {
-      const wrapper = createWrapper();
-      expect(wrapper.state().modalDisplayed).toBe(false);
+    it('calls the openShareDialog and handleClose props when clicked', () => {
+      const handleClose = jest.fn();
+      const openDownloadDialog = jest.fn();
+      const wrapper = createWrapper({ handleClose, openDownloadDialog });
       wrapper.find('WithStyles(MenuItem)').simulate('click');
-      expect(wrapper.state().modalDisplayed).toBe(true);
-    });
-  });
-
-  describe('Dialog', () => {
-    it('renders a dialog that is open/closed based on the component state', () => {
-      const wrapper = createWrapper();
-      expect(wrapper.state().modalDisplayed).toBe(false);
-      expect(wrapper.find('WithStyles(Dialog)').props().open).toBe(false);
-      wrapper.setState({ modalDisplayed: true });
-      expect(wrapper.find('WithStyles(Dialog)').props().open).toBe(true);
-    });
-
-    it('renders a CanvasDownloadLinks componewnt for every canvas', () => {
-      const mockCanvas = id => ({
-        id,
-        getHeight: () => 4000,
-        getWidth: () => 1000,
-        getCanonicalImageUri: () => 'https://example.com/iiif/abc123/full/9000,/0/default.jpg',
-      });
-      const wrapper = createWrapper({ canvases: [mockCanvas('abc123'), mockCanvas('xyz321')] });
-      expect(wrapper.find('CanvasDownloadLinks').length).toBe(2);
-    });
-
-    it('has a close button that updates the modealDisplay state to false', () => {
-      const wrapper = createWrapper();
-      wrapper.setState({ modalDisplayed: true });
-      expect(wrapper.state().modalDisplayed).toBe(true);
-      wrapper.find('WithStyles(Button)').simulate('click');
-      expect(wrapper.state().modalDisplayed).toBe(false);
-    });
-
-    describe('ManifestDownloadLinks', () => {
-      it('is not rendered if hte manifest has no renderings', () => {
-        const wrapper = createWrapper();
-
-        expect(wrapper.find('ManifestDownloadLinks').length).toBe(0);
-      });
-      it('rendered if the manifest has renderings', () => {
-        const rendering = { id: '', getLabel: () => {}, getFormat: () => {} };
-        const wrapper = createWrapper({
-          manifest: {
-            getSequences: () => [
-              { getRenderings: () => [rendering] },
-            ],
-          },
-        });
-
-        expect(wrapper.find('ManifestDownloadLinks').length).toBe(1);
-      });
+      expect(handleClose).toHaveBeenCalled();
+      expect(openDownloadDialog).toHaveBeenCalled();
     });
   });
 });
